# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your API Title",
        "description": "Your API description",
        "contact": {
            "email": "contact@example.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/user-preference/auth": {
            "get": {
                "tags": [
                    "User Preferences"
                ],
                "summary": "Get user preferences by Authentication",
                "description": "Retrieve user preferences for the authenticated user.",
                "operationId": "getUserPreferencesByAuth",
                "responses": {
                    "200": {
                        "description": "User preferences retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user_preferences": {
                                            "$ref": "#/components/schemas/UserPreference"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User preferences not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "User Preferences"
                ],
                "summary": "Update user preferences by Authentication",
                "description": "Update or create user preferences for the authenticated user.",
                "operationId": "updateUserPreferencesByAuth",
                "requestBody": {
                    "description": "User preferences to be updated or created",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "preferred_sources",
                                    "preferred_categories",
                                    "preferred_authors"
                                ],
                                "properties": {
                                    "preferred_sources": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    },
                                    "preferred_categories": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    },
                                    "preferred_authors": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User preferences updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user_preferences": {
                                            "$ref": "#/components/schemas/UserPreference"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "User preferences created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user_preferences": {
                                            "$ref": "#/components/schemas/UserPreference"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "UserPreference": {
                "title": "User Preference",
                "description": "User preference model",
                "properties": {
                    "id": {
                        "description": "The unique identifier for the user preference",
                        "type": "integer"
                    },
                    "user_id": {
                        "description": "The ID of the user associated with the preference",
                        "type": "integer"
                    },
                    "preferred_sources": {
                        "description": "Preferred sources",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "preferred_categories": {
                        "description": "Preferred categories",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "preferred_authors": {
                        "description": "Preferred authors",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "created_at": {
                        "description": "The date and time when the user preference was created",
                        "type": "string",
                        "example": "2022-05-10 12:00:00"
                    },
                    "updated_at": {
                        "description": "The date and time when the user preference was last updated",
                        "type": "string",
                        "example": "2022-05-10 12:00:00"
                    }
                },
                "type": "object"
            }
        }
    },
...
}